requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
requests>=2.31.0
Pillow>=10.0.0
anthropic>=0.25.0
beautifulsoup4>=4.12.0

